# swagger.json
{
	"swagger": "2.0",
	"info": {
		"version": "v1",
		"title": "Price.WebApi"
	},
	"host": "localhost:52620",
	"schemes": ["http"],
	"paths": {
...
		"/api/xpath": {
			"get": {
				"tags": ["Xpath"],
				"summary": "",
				"operationId": "Xpath_XpathGet",
				"consumes": [],
				"produces": ["application/json",
				"text/json",
				"application/xml",
				"text/xml"],
				"parameters": [{
					"name": "link",
					"in": "query",
					"description": "",
					"required": true,
					"type": "string"
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/XPathDto"
						}
					}
				}
			},
			"post": {
				"tags": ["Xpath"],
				"summary": "",
				"operationId": "Xpath_XpathPost",
				"consumes": ["application/json",
				"text/json",
				"application/xml",
				"text/xml",
				"application/x-www-form-urlencoded"],
				"produces": ["application/json",
				"text/json",
				"application/xml",
				"text/xml"],
				"parameters": [{
					"name": "dto",
					"in": "body",
					"description": "",
					"required": true,
					"schema": {
						"$ref": "#/definitions/XPathDto"
					}
				}],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"format": "int32",
							"enum": [100,
							101,
							200,
							201,
							202,
							203,
							204,
							205,
							206,
							300,
							300,
							301,
							301,
							302,
							302,
							303,
							303,
							304,
							305,
							306,
							307,
							307,
							400,
							401,
							402,
							403,
							404,
							405,
							406,
							407,
							408,
							409,
							410,
							411,
							412,
							413,
							414,
							415,
							416,
							417,
							426,
							500,
							501,
							502,
							503,
							504,
							505],
							"type": "integer"
						}
					}
				}
			}
		}
	},
	"definitions": {
...
		"XPathDto": {
			"description": "",
			"type": "object",
			"properties": {
				"XPathName": {
					"description": "",
					"type": "string"
				},
				"XPathPrice": {
					"description": "",
					"type": "string"
				},
				"uri": {
					"description": "",
					"type": "string"
				},
				"domain": {
					"description": "",
					"type": "string"
				},
				"idndomain": {
					"description": "",
					"type": "string"
				},
				"name": {
					"description": "",
					"type": "string"
				},
				"Price": {
					"description": "",
					"type": "string"
				},
				"collectedAt": {
					"format": "int64",
					"description": "",
					"type": "integer"
				},
				"id": {
					"description": "",
					"type": "string"
				}
			}
		}
	}
}